routing: Add a container with a custom router OS based on dnsmasq.
This is required to bridge hardware devices like SIP phone with the
other services provided or planned by Reclaimail. It is alpha at best.

It will require some dnsmasq patches to add some missing features before
this is generic enough. Right now it hardcodes some interface name in
the config because Lua can't set it. It /is/ possible to hack this by
using stdin to change the config, but this has drawbacks.