routing: Patch dnsmasq to have a more useful Lua API.
This is currently a work in progress, but it replaces the config
file with Lua code and adds a partial higher level wrapper around
the dnsmasq config syntax.

The API will change a lot, but this already works to configure the
interfaces and make PXE work with many ISO files. For now there is
still hardcoded mac address and paths.

The dockerfile currently doesn't compile the patched dnsmasq, this
will come later.