routing: Sync with my private repo.
This code is from Dec 2020.

It adds a lot of Lua code to further advance the long term goal of
using this OS to provision PXE devices like PCs, server, phones
and rPis.

The patched `dnsmasq` has some design bugs. For example, `dnsmasq`
runs the scripts in a sub-process, but the patch breaks it and run
code in the main process. That was done because adding enough IPC
dance to implement the PXE request interception across the 2
process was taking too long. Obviously it means the patch cannot
be upstreamed because they are a denial of service attack waiting
to happen.

That being said, the with this code, the vision is starting become
a reality. It *can* boot PCs and rPis over PXE and it almost works
ok-ish.

This commit doesn't contain the `ganesha` based NFS server
container. That container is actually needed to boot real ISOs.
Otherwise they have to boot into ram over tFTP, which is slow,
error prone and slow down dnsmasq to a point it's unusable.