phone: Merge the long delayed SIP server components.
Started in 2018, never got to the quality to dog-food them until now.

This is a minimally viable product. It works ok for basic use cases,
but is missing half of the advanced features and what isn't tested
is very unlikely to work. It has no documentation and no tests (yet).

You can receive and send calls. Voicemails also work ok-ish.